server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Justin Chapman", email = "commonmodestudio@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "python-ulid[pydantic]>=3.0.0",
]

[project.scripts]
server = "server:main"

[tool.scikit-build]
minimum-version = "build-system.requires"
build-dir = "build/{wheel_tag}"

[build-system]
requires = ["scikit-build-core>=0.10", "pybind11"]
build-backend = "scikit_build_core.build"

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pybind11>=2.13.6",
    "pytest>=8.3.4",
    "ruff>=0.9.2",
]
